Bug fixes
5. 05/02/2014 - Fixed issue with controlservices not looping through
folders properly, Thanks to Anthony Pagan and Eric Moody for finding
this!
6. 05/02/2014 - Fixed issue with truncateappend requiring a datefield
when it should be optional, Thanks to Anthony Pagan and Eric Moody for
finding this!
7. 05/02/2014 - Fixed issue with email function, error was being thrown
when starttls was envoked, added some error trapping and user was then
able to send emails, Thanks to Jeff Trimmer for finding this!